Chore: Reorganise to MVC Folder Structure
