+++ src/main/java/marsmission/marsrover/MissionControl.java
@@ -1,4 +1,4 @@
-package marsmission.marsrover.app;
+package marsmission.marsrover;
 
 import marsmission.marsrover.view.ConsoleMessenger;
